requirements.txt
django==1.7

markdown
pytils
django-braces==1.4.0
django-bower==5.0.1
django-compressor==1.4
django-model-utils
django-guardian

psycopg2

pillow
django-imagekit

djangorestframework
django-filter
